Move examples and update API
