+++ README.md
@@ -39,7 +39,7 @@ from sklearn.datasets import load_breast_cancer
 from sklearn.metrics import accuracy_score
 from sklearn.model_selection import train_test_split
 
-from contexttab.contexttab import ConTextTabClassifier
+from contexttab import ConTextTabClassifier
 
 # Load sample data
 X, y = load_breast_cancer(return_X_y=True)
@@ -63,7 +63,7 @@ from sklearn.datasets import fetch_openml
 from sklearn.metrics import r2_score
 from sklearn.model_selection import train_test_split
 
-from contexttab.contexttab import ConTextTabRegressor
+from contexttab import ConTextTabRegressor
 
 
 # Load sample data
