+++ contexttab/__init__.py
@@ -1,0 +1,3 @@
+from contexttab.contexttab import ConTextTabClassifier, ConTextTabRegressor
+
+__all__ = ['ConTextTabClassifier', 'ConTextTabRegressor']
